page num field changed

+++ books/models.py
@@ -6,7 +6,7 @@ class Book(models.Model):
     autor = models.CharField(max_length=30)
     data_publikacji = models.DateField(help_text='yyyy-mm-dd')
     numer_isbn = models.CharField(max_length=20, verbose_name='Numer ISBN')
-    liczba_stron = models.IntegerField(default=0)
+    liczba_stron = models.PositiveSmallIntegerField(default=0)
     link_do_okładki = models.URLField()
     język_publikacji = models.CharField(max_length=20)
